Move member & path stuff into seperate module
